Update k0s config if user makes a change
The current check is only to see if the file exists, if it does then `k0s_config` is pretty much
skipped. If a user wishes to make change, they that to set `force` option. But this either means
leaving `force` to `true` or making multiple changes.

This commit, will check if there are any user supplied changes, and if there is, then write out the
Yaml to k0s config file